Use float for AQI

+++ main.py
@@ -131,11 +131,11 @@ def upload_measurement():
 
     aqi_value = None
     if "SDS_P1" in data_points and "SDS_P2" in data_points:
-        aqi_value = int(
+        aqi_value = float(
             aqi.to_aqi([(aqi.POLLUTANT_PM10, data_points["SDS_P1"]), (aqi.POLLUTANT_PM25, data_points["SDS_P2"])])
         )
     elif "PMS_P1" in data_points and "PMS_P2" in data_points:
-        aqi_value = int(
+        aqi_value = float(
             aqi.to_aqi([(aqi.POLLUTANT_PM10, data_points["PMS_P1"]), (aqi.POLLUTANT_PM25, data_points["PMS_P2"])])
         )
     else:
